Added Exception implementation in Graphics class

+++ D3D11Tutorial/src/App/App.cpp
@@ -31,7 +31,7 @@ int App::Run()
 
 void App::OnFrame()
 {
-    const float c = std::sin(m_Timer.Peek()) / 2.0f + 0.5;
+    const float c = (float)(std::sin(m_Timer.Peek()) / 2.0 + 0.5);
     m_Window.GetGraphics().ClearBuffer(c, c, 1.0f);
     m_Window.GetGraphics().EndFrame();
 }
